Refactor of query modal complete except custom query click
