+++ frontend/pages/hosts/HostDetailsPage/HostDetailsPage.tsx
@@ -51,7 +51,7 @@ import {
   secondsToHms,
 } from "fleet/helpers";
 // @ts-ignore
-import SelectQueryModal from "./SelectQueryModalTsx";
+import SelectQueryModal from "./SelectQueryModal";
 import TransferHostModal from "./TransferHostModal";
 import PolicyDetailsModal from "./HostPoliciesTable/PolicyDetailsModal";
 import {
