+++ frontend/pages/hosts/HostDetailsPage/SelectQueryModal/_styles.scss
