+++ frontend/pages/hosts/HostDetailsPage/SelectQueryModal/index.ts
